Refactor deletions (#181)
### Description

<!-- Provide a comprehensive description here about what your PR aims to
solve. -->

<!-- You may also add additional context -->

Increase data integrity by returning the deleted entity / a boolean for
each deleted entity in the repository layer.

This fixes a data integrity issue we had when the repository would not
tell the service if the entity was actually deleted in some cases.

Closes #137 

---

### PR Checklist

<!-- Please do not remove this section -->

<!-- Mark each item with an "x" ([ ] becomes [x]) -->

- [x] Read the Developer's Guide in
[CONTRIBUTING.md](https://github.com/agnyz/bedstack/blob/main/CONTRIBUTING.md)
- [x] Use a concise title to represent the changes introduced in this PR
- [x] Provide a detailed description of the changes introduced in this
PR, and, if necessary, some screenshots
- [x] Reference an issue or discussion where the feature or changes have
been previously discussed
- [x] Add a failing test that passes with the changes introduced in this
PR, or explain why it's not feasible
- [x] Add documentation for the feature or changes introduced in this PR
to the docs; you can run them with `bun docs`


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Bug Fixes**
- Improved reliability of article and comment deletion by providing
clear success or failure feedback and explicit error handling if
deletion fails.
- Enhanced unfollow functionality to clearly indicate whether the
unfollow action was successful.

- **Refactor**
- Updated several deletion-related methods to return boolean values,
making their outcomes more transparent to users.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->